feat: 텍스트박스 이중팝업 해결
