+++ frontend/fr-study/src/pages/Mainpage/mainpage.css
