+++ frontend/fr-study/src/pages/Mainpage/popup.css
@@ -33,7 +33,7 @@
   height: 50px;
 }
 
-.close-button {
+.popup-close-button {
   background: none;
   border: none;
   font-size: 24px;
